+++ lib/pages/verse_browser/verse_browser_manager.dart
@@ -31,7 +31,6 @@ class VerseBrowserManager {
   }
 
   void onFinishedEditing(String? verseId) async {
-    if (verseId == null) return;
     init(_collectionId);
   }